+++ aksel.nav.no/website/pages/eksempler/button/secondary-neutral.tsx
@@ -1,0 +1,17 @@
+import { Button } from "@navikt/ds-react";
+import { withDsExample } from "components/website-modules/examples/withDsExample";
+
+const Example = () => {
+  return <Button variant="secondary-neutral">Secondary</Button>;
+};
+
+export default withDsExample(Example);
+
+/* Storybook story */
+export const Demo = {
+  render: Example,
+};
+
+export const args = {
+  index: 12,
+};