+++ aksel.nav.no/website/pages/produktbloggen/[slug].tsx
@@ -13,7 +13,7 @@ import TableOfContents from "components/website-modules/TOC";
 import { getImage } from "components/website-modules/utils/get-image";
 import { PreviewSuspense } from "next-sanity/preview";
 import Head from "next/head";
-import Image from "next/image";
+import Image from "next/legacy/image";
 import { GetServerSideProps } from "next/types";
 import { lazy } from "react";
 import NotFotfund from "../404";
@@ -88,14 +88,14 @@ const Page = ({
       <main
         tabIndex={-1}
         id="hovedinnhold"
-        className="aksel-artikkel xs:pb-32 #FEFCE9 overflow-hidden bg-[#FEFCE9] pt-[8vw] pb-16 focus:outline-none"
+        className="aksel-artikkel #FEFCE9 overflow-hidden bg-[#FEFCE9] pt-[8vw] pb-16 focus:outline-none sm:pb-32"
       >
         <div className="px-4">
           <div className="dynamic-wrapper-prose text-center">
             <Heading
               level="1"
               size="xlarge"
-              className="algolia-index-lvl1 hyphen text-deepblue-800 mt-1 break-words text-5xl leading-[1.2]"
+              className="hyphen text-deepblue-800 mt-1 break-words text-5xl leading-[1.2]"
             >
               {blogg.heading}
             </Heading>
@@ -185,7 +185,7 @@ const Page = ({
             )}
             <BodyShort
               as="span"
-              className="text-text-subtle flex justify-center whitespace-nowrap"
+              className="text-text-subtle flex justify-center"
             >
               Publisert: {dateStr(blogg?.publishedAt ?? blogg?._updatedAt)}
             </BodyShort>
@@ -195,7 +195,7 @@ const Page = ({
               <Heading level="2" size="large">
                 Flere blogginnlegg
               </Heading>
-              <ul className="mt-12 grid gap-x-6 gap-y-6 sm:grid-cols-2 sm:gap-y-10 md:gap-x-6 lg:grid-cols-3">
+              <ul className="mt-12 grid gap-x-6 gap-y-6 md:grid-cols-2 md:gap-y-10 md:gap-x-6 lg:grid-cols-3">
                 {morePosts.map((post) => (
                   <BloggCard key={post._id} blog={post} />
                 ))}